accounts-db: Sampled LRU eviction
Instead of rigid LRU eviction, perform sampled LRU eviction.

Sampled LRU eviction takes K random keys and picks the one with the
lowest update timestamp. This way, even though the evicted element is
not always the oldest in the whole cache, removes the necessity of
maintaining a queue and reduces the contention caused by locking the
queue.

Add a benchmark which measures time spent on concurrent writes and reads
on a full cache (where each write causes eviction).

The result before this change, for 128 read and write threads, is:

```
cache_eviction/cache_eviction/128
                        time:   [20.737 s 20.936 s 21.138 s]
```

With sampled LRU eviction (K=16), it improves to:

```
cache_eviction/cache_eviction_k_16/128
                        time:   [7.6609 s 7.7177 s 7.7847 s]
```