accounts-db: Improve sharding of the read-only cache
The default amount of shards in `DashMap` is `num_cpus * 4`. That means
256 on validators with 64 threads. A number of accounts held in caches
on mainnet beta validators with default configuration is around 44-56K,
which results in around 200 accounts per shard. That means, locking a
shard locks an access to 200 accounts.

Fix that by increasing the amount of shards to 65536, which in the best
case  will keep just one account per shard. Acquiring a lock shouldn't
lock access to any other accounts.

A single `RwLock<RawTable<(K, V)>>`, without the key and value included,
takes 48 bytes. Therefore, 65536 shards are taking 3145728 B, which
rounds up to 3MB. It's an acceptable increase in memopry usage.