core/tpu: Add zero-copy packet implementations (`TpuPacket`)
Add new packet implementations, based on `bytes::Bytes`[0] and
`bytes::BytesMut`[1], which provide a zero-copy abstraction, where
calling `clone()` doesn't make an actual copy of the underlying data
and all instances point to the same memory.

* `TpuPacketMut` - a mutable packet buffer with metadata, used by the
  QUIC servers.
* `TpuPacket` - an immubable packet buffer with metadata, used through
  all TPU stages.

To keep the compatibility of sigverify with both `Packet` and
`TpuPacket`, introduce the `PacketRead` trait.

[0] https://docs.rs/bytes/1.10.1/bytes/struct.Bytes.html
[1] https://docs.rs/bytes/1.10.1/bytes/struct.BytesMut.html