core/tpu: Add `BytesPacket`, a zero-copy packet implementation
The `Packet` struct stores the payload in a byte array. QUIC servers,
based on quinn, provide received data into `bytes::Bytes`[0]. The usage
of `Packet` across the TPU code requires a copy from `Bytes` to
`Packet`.

That defeats the biggest advantage of `Bytes` - its zero-copy
abstraction, where calling `clone()` doesn't make an actual copy of the
underyling packet data and all instances point to the same memory.

Fix that by providing a new packet type - `BytesPacket` - which uses
`Bytes` for storing the payload. That new type of packet is used in
QUIC servers of TPU.

Rename `PacketBatch` to `PinnedPacketBatch`. Add a new
`BytesPacketBatch` type, which is a vector of `BytesPacket`.

To make sigverify work with both types of packets (`Packet` and
`BytesPacket`), provide the abstraction:

* `PacketBatch` enum, carrying `PinnedPacketBatch` and
  `BytesPacketBatch` as variants. That enum proxies all the methods
  of both bath types.
* `PacketRef` and `PacketRefMut` enums, serving as wrapper types
  carrying references to an underlying packet instance.
* `PacketBatchIter` and `PacketBatchIterMut` enums, which implement
  `Iterator` and `ParallelIterator` and allow an uniform way of
  iteration over `PacketBatch`, regardless of the underlying variant.

CUDA sigverify still requires `Packet` and `PinnedPacketBatch`. The
conversion of all packet batches to `PinnedPacketBatch` is made right
before calling CUDA. In case of a `PacketBatch` being `Bytes`-based,
that means making one copy.

[0] https://docs.rs/bytes/1.10.1/bytes/struct.Bytes.html