runtime: Avoid locking during stake vote rewards calculation
`calculate_stake_vote_rewards` was storing accumulated rewards per vote
account in a `DashMap`, which then was used in a parallel iterator over
all stake delegations. That resulted in high contention on the `DashMap`
locks and rayon threads being busy with acquiring a lock of a vote
account shard.

Remove necessity of using `DashMap` by providing a custom thread pool
implementation, which allows to provide worker states - custom data,
which is local for each thread in the pool and therefore doesn't
require any locking. Worker states are passed as a mutable slice to
the thread pool scope. The scope is dropped only once all threads are
done. After that, the worker states can be used by the main thread
again.