runtime: Avoid locking during stake vote rewards calculation
`calculate_stake_vote_rewards` was storing accumulated rewards per vote
account in a `DashMap`, which then was used in a parallel iterator over
all stake delegations. That resulted in high contention on the `DashMap`
locks and rayon threads being busy with acquiring a lock of a vote
account shard.

The time spent on these calculations was ~208.47ms:

```
redeem_rewards_us=208475i
```

Remove necessity of using `DashMap` by providing a custom thread pool
implementation, which allows to provide worker states - custom data,
which is local for each thread in the pool and therefore doesn't
require any locking. Worker states are passed as a mutable slice to
the thread pool scope. The scope is dropped only once all threads are
done. After that, the worker states can be used by the main thread
again.

The profile indicates that the new code spends:

* 27% of time on Vec::push. Even though the code already pre-allocates
  the vectors with a heuristic size
  (`stake_delegations.len().div_ceil(num_workers)`), some threads end up
  processing more stake delegations than others.
* 37% on the actual calculations in
  `solana_runtime::inflation_rewards::redeem_rewards`.
* 14% on adding elements to hash map. Unfortunately, we can't predict
  its size beforehand.

The time of calculation goes down to ~16.50ms:

```
redeem_rewards_us=16501i
```