runtime: Avoid locking during stake vote rewards calculation
`calculate_stake_vote_rewards` was storing accumulated rewards per vote
account in a `DashMap`, which then was used in a parallel iterator over
all stake delegations. That resulted in high contention on the `DashMap`
locks and rayon threads being busy with acquiring a lock of a vote
account shard.

The time spent on these calculations was ~208.47ms:

```
redeem_rewards_us=208475i
```