runtime: Avoid redundant collections of stake delegations into a vector
Stake delegations are stored in an `im::HashMap`, which is convenient
to use as a per-bank cache, because of its CoW properties:

* `clone()` is just copying a pointer and increasing a reference counter
* post-clone modification of a single node, makes a copy only of that
  node, not of the whole map

However, `im::HashMap`:

* has terrible traversal performance - a full iteration over a map with
  1,000,000 elements takes 200ms
* doesn't support parallel iteration with Rayon
  * using `par_bridge` performs even worse than a single-threaded
    iteration - it takes 1.3s

Therefore, it's better to `collect()` the `im::HashMap` before
calculating rewards.

The problem is that there were two `collect()` calls made during one
epoch boundary.

Fix that by collecting only once and passing the vector of delegations
along.