runtime: Add `Stakes::stake_delegations_vec` method
Stake delegations are stored as hash array mapped trie (HAMT)[0], which
means that inserts, deletions and lookups are average-case O(1) and
worst-case O(log n). However, the performance of iterations is poor due
to depth-first traversal and jumps. Currently it's also impossible to
iterate over it with rayon.

That issue is known and handled by converting the HAMT to a vector with
`stakes.stake_delegations.iter().collect()`. Move that trick to a
dedicated method that describes the performance consequences.

[0] https://en.wikipedia.org/wiki/Hash_array_mapped_trie