runtime: Add `FilteredStakeDelegation` wrapper type
`filter_stake_delegations` was always collecting the stake delegations.

To allow re-using already existing vectors or slices of stake
delegations, add `FilteredStakeDelegation` type that wraps a `Cow` of
stake delegations and provides a parallel iterator that yields elements,
filtering them by the minimum stake.

Note that the wrapper yields an `Option`. That makes it possible to know
the size of the iterator without collecting elements. That property is
critical for the ability to pre-allocate data structures that contain
processed stake delegations. However, it adds a necessity of handling
`None` elements while iterating.