runtime: Move `store_vote_accounts_partitioned` out
The previous change was about taking immutable `&self` in
`Bank::begin_partitioned_rewards`. However, `store_vote_accounts_partitioned`,
even though it does not take a mutable reference on `Bank`, acquires
a write lock on the status cache. That would lead to a deadlock if
we acquire a read lock for the span of `begin_partitioned_rewards`.

It's safer if we move `store_vore_accounts_partitioned` out to ...,
making it clear that it will acquire a write lock.