zk-token-sdk: Do not require passing owned vectors to `RangeProof`
The current implementation of `RangeProof::new` and `RangeProof::verify`
requires the caller to pass inputs as owned vectors. That requires heap
allocations even if the number of inputs is small (even as small as 1).

Remove that requirement by accepting `AsRef<[T]>` trait as inputs, which
still makes it possible to pass `Vec<T>`, so we don't break the old
users relying on the old API. But for callers who want performance, it
accepts also the following types that can be allocated on the stack:

- `[T]`
- `[T; N]`
- `&[T]`

Furthermore, single elements can be passed using `slice::from_ref`[0],
e.g.

```rust
let proof = RangeProof::new(
    slice::from_ref(&55),
    slice::from_ref(&64),
    slice::from_ref(&open),
    &mut transcript_create,
)
```

which does not perform any allocation at all. Instead, it performs raw
pointer casting to coerce a reference to a single-element slice[1][2].

Additionally, `AsRef<T>` makes it possible to pass the following
iterable types:

- `std::slice::Iter<'_, T>`
- `IterMut<'_, T>`
- `IntoIter<T, A>`